+++ src/main/java/cart/application/service/product/ProductWriteService.java
@@ -1,7 +1,7 @@
 package cart.application.service.product;
 
 import cart.application.repository.product.ProductRepository;
-import cart.application.service.product.dto.ProductCreateDto;
+import cart.application.service.product.dto.ProductCommandDto;
 import cart.domain.product.Product;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,16 +18,16 @@ public class ProductWriteService {
         this.productRepository = productRepository;
     }
 
-    public Long createProduct(final ProductCreateDto productCreateDto) {
-        final Product product = new Product(productCreateDto);
+    public Long createProduct(final ProductCommandDto productCommandDto) {
+        final Product product = new Product(productCommandDto);
         return productRepository.createProduct(product);
     }
 
-    public void updateProduct(final Long productId, final ProductCreateDto productCreateDto) {
+    public void updateProduct(final Long productId, final ProductCommandDto productCommandDto) {
         productRepository.findById(productId)
                 .orElseThrow(() -> new NoSuchElementException("존재하지 않는 상품입니다."));
 
-        final Product product = new Product(productId, productCreateDto);
+        final Product product = new Product(productId, productCommandDto);
         productRepository.updateProduct(product);
     }
