+++ src/main/java/cart/ui/cartitem/CartItemController.java
@@ -1,11 +1,18 @@
 package cart.ui.cartitem;
 
+import cart.application.service.cartitem.CartItemReadService;
 import cart.application.service.cartitem.CartItemWriteService;
 import cart.application.service.cartitem.dto.CartItemCreateDto;
 import cart.application.service.cartitem.dto.CartItemUpdateDto;
+import cart.application.service.cartitem.dto.CartResultDto;
+import cart.application.service.order.OrderMakeService;
+import cart.application.service.order.dto.CreateOrderDto;
 import cart.domain.member.Member;
 import cart.ui.cartitem.dto.CartItemQuantityUpdateRequest;
 import cart.ui.cartitem.dto.CartItemRequest;
+import cart.ui.cartitem.dto.CartPaymentResponse;
+import cart.ui.cartitem.dto.CartResponse;
+import cart.ui.order.dto.request.CreateOrderRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,12 +20,34 @@ import java.net.URI;
 
 @RestController
 @RequestMapping("/cart-items")
-public class CartItemWriteController {
+public class CartItemController {
 
+    private final CartItemReadService cartItemReadService;
+    private final OrderMakeService orderMakeService;
     private final CartItemWriteService cartItemWriteService;
 
-    public CartItemWriteController(final CartItemWriteService cartItemWriteService) {
+    public CartItemController(
+            final CartItemReadService cartItemReadService,
+            final OrderMakeService orderMakeService,
+            final CartItemWriteService cartItemWriteService) {
+        this.cartItemReadService = cartItemReadService;
+        this.orderMakeService = orderMakeService;
         this.cartItemWriteService = cartItemWriteService;
+    }
+
+    @GetMapping
+    public ResponseEntity<CartResponse> showCartItems(final Member memberAuth) {
+        final CartResultDto cartResult = cartItemReadService.findByMember(memberAuth);
+        return ResponseEntity.ok(CartResponse.from(cartResult));
+    }
+
+    @PostMapping("/payment")
+    public ResponseEntity<CartPaymentResponse> calculatePrice(
+            final Member member,
+            @RequestBody CreateOrderRequest createOrderRequest
+    ) {
+        final int paymentPrice = orderMakeService.getPaymentPrice(member, CreateOrderDto.from(createOrderRequest));
+        return ResponseEntity.ok(new CartPaymentResponse(paymentPrice));
     }
 
     @PostMapping
@@ -47,4 +76,5 @@ public class CartItemWriteController {
 
         return ResponseEntity.noContent().build();
     }
+
 }
